% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sillylogic_TRUE.R
\docType{data}
\name{sillyTRUE}
\alias{COUNTMEIN}
\alias{HELLYES}
\alias{WHYNOT}
\alias{YES}
\alias{YESPLEASE}
\alias{sillyTRUE}
\title{Fun Alternate Values for TRUE}
\format{\preformatted{ logi TRUE
}}
\source{
https://twitter.com/michaelhoffman/status/630420081042849792
}
\usage{
YES

YESPLEASE

HELLYES

WHYNOT

COUNTMEIN
}
\description{
Inspired by a Twitter post (see Sources), provides alternate values
  that may be used in place of the traditional logical value \code{TRUE}.
}
\examples{
x <- data.frame(letter = letters[1:10],
                stringsAsFactors = HELLYES)
class(x$letter)
}
\seealso{
\code{\link{sillyFALSE}}, \code{\link{surprise_me}}
}
\keyword{datasets}

