% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/surprise_me.R
\name{surpise_me}
\alias{surpise_me}
\alias{surprise_me}
\title{Random values for TRUE and FALSE}
\source{
https://twitter.com/michaelhoffman/status/630420081042849792
}
\usage{
surprise_me(n = 1)
}
\arguments{
\item{n}{desired length of the returned vector}
}
\description{
If you like to live dangerously, randomly choose your booleans
  and get different results every time.
}
\seealso{
\code{\link{sillyFALSE}}, \code{\link{sillyTRUE}}
}

